ebdjango/requirements.txt
absl-py>=0.6.0
astor>=0.7.1
awscli>=1.16.57
backcall>=0.1.0
bleach>=3.0.2
botocore>=1.12.47
colorama>=0.3.9
cycler>=0.10.0
decorator>=4.3.0
defusedxml>=0.5.0
Django>=2.1.4
djangorestframework>=3.10.3
docutils>=0.14
entrypoints>=0.2.3
gast>=0.2.0
grpcio>=1.16.0
h5py>=2.8.0
ipykernel>=5.1.0
ipython>=7.0.1
ipython-genutils>=0.2.0
ipywidgets>=7.4.2
jedi>=0.13.1
Jinja2>=2.10
jmespath>=0.9.3
jsonschema>=2.6.0
jupyter>=1.0.0
jupyter-client>=5.2.3
jupyter-console>=6.0.0
jupyter-core>=4.4.0
kiwisolver>=1.0.1
Markdown>=3.0.1
MarkupSafe>=1.0
matplotlib>=3.0.0
mistune>=0.8.4
nbconvert>=5.4.0
nbformat>=4.4.0
notebook>=5.7.0
numpy>=1.15.3
pandas>=0.23.4
pandocfilters>=1.4.2
parso>=0.3.1
pickleshare>=0.7.5
prometheus-client>=0.4.2
prompt-toolkit>=2.0.6
protobuf>=3.6.1
psycopg2>=2.7.5
pyasn1>=0.4.4
Pygments>=2.2.0
pyparsing>=2.2.2
python-dateutil>=2.7.5
pytz>=2018.6
pywinpty>=0.5.4
PyYAML>=3.13
pyzmq>=17.1.2
qtconsole>=4.4.2
rsa>=3.4.2
s3transfer>=0.1.13
scikit-learn>=0.20.0
scipy>=1.1.0
Send2Trash>=1.5.0
simplegeneric>=0.8.1
six>=1.11.0
termcolor>=1.1.0
terminado>=0.8.1
testpath>=0.4.2
tornado>=5.1.1
traitlets>=4.3.2
urllib3>=1.24.1
wcwidth>=0.1.7
webencodings>=0.5.1
Werkzeug>=0.14.1
widgetsnbextension>=3.4.2
